refactor: Unify types — use src/schemas/pipeline.ts as canonical source
- Expanded pipeline.ts with TaskStatusSchema, CheckpointSchema, and complete SessionStateSchema
- SessionState now includes all 15 fields needed by Role 2.2: session_id, version, goal, current_task, completed_tasks, key_decisions, active_files, tasks, summaries, artifacts, metadata, last_summary, next_action, updated_at
- Task schema now includes status, input_hash, output_summary fields
- src/types/state.ts now re-exports canonical schemas/types from pipeline.ts (removed 106 lines of duplication)
- src/types/index.ts exports pipeline.ts schemas as the primary source
- SessionStateManager already uses correct imports via src/types/index.ts re-exports
- TypeScript compilation verified
