+++ src/types/index.ts
@@ -1,2 +1,6 @@
+// Core schemas and types (canonical source: src/schemas/pipeline.ts)
+export * from '../schemas/pipeline.js';
+
+// Role 2.2 specific extensions and re-exports
 export * from './state.js';
 export * from './context.js';
